Add ZeroDivisionError and exit()
This commit introduces a ZeroDivisionError exception and an `exit()`
function to the Hython interpreter. The `exit()` function allows scripts
to terminate execution with a specified exit code.

Additionally, the division and assignment division operators now check
for division by zero and raise the appropriate exception. The README has
also been slightly updated for clarity.